Avoid `ledc_hal_timer_rst()` when frequency hasn't changed.
This "fixes" a bug where phase shifts whenever the duty cycle is
changed, and prevents a runt pulse whever duty cycle is set to zero.

This is not an ideal fix. I'd prefer to move all initialization into the
the diver's init function, but this  isn't currently possible because the
clock selection is dynamic based on the requested period.

Signed-off-by: Braiden Kindt <braiden@kindt.dev>